third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-playback-multiple-sessions.html

<!DOCTYPE html>
<html>
    <head>
        <title>Clear Key Playback with Multiple Sessions</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <video id="testVideo"></video>
        <div id="log"></div>
        <p>Test playback of encrypted media using Clear Key key system with multiple sessions.</p>
        <script>
            async_test(function(test)
            {
                var video = document.getElementById('testVideo');
                var audioMediaKeySession = null;
                var videoMediaKeySession = null;
                var audioKeyProvided = false;
                var videoKeyProvided = false;

                // The 2 streams use different key ids and different keys.
                var audioKeyId = '1234567890123456';
                var audioKey = new Uint8Array([0x30, 0x30, 0x62, 0xF1, 0x68, 0x14, 0xD2, 0x7B,
                                               0x68, 0xEF, 0x12, 0x2A, 0xFC, 0xE4, 0xAE, 0x0A]);
                var videoKeyId = '0123456789012345';
                var videoKey = new Uint8Array([0x7A, 0x7A, 0x62, 0xF1, 0x68, 0x14, 0xD2, 0x7B,
                                               0x68, 0xEF, 0x12, 0x2A, 0xFC, 0xE4, 0xAE, 0x0A]);

                function onNeedKey(event)
                {
                    var keyId = String.fromCharCode.apply(null, event.initData);

                    var mediaKeySession = video.mediaKeys.createSession();
                    waitForEventAndRunStep('message', mediaKeySession, onMessage, test);
                    mediaKeySession.generateRequest(event.contentType, event.initData).then(function() {
                        if (keyId == videoKeyId) {
                            assert_equals(videoMediaKeySession, null);
                            videoMediaKeySession = mediaKeySession;
                        } else {
                            assert_equals(keyId, audioKeyId);
                            assert_equals(audioMediaKeySession, null);
                            audioMediaKeySession = mediaKeySession;
                        }
                    }).catch(function(error) {
                        forceTestFailureFromPromise(test, error);
                    });
                }

                function onMessage(event)
                {
                    var keyId = extractSingleKeyIdFromMessage(event.message);
                    if (event.target == videoMediaKeySession) {
                        assert_equals(String.fromCharCode.apply(null, keyId), videoKeyId);
                        var jwkSet = stringToUint8Array(createJWKSet(createJWK(keyId, videoKey)));
                        videoMediaKeySession.update(jwkSet).then(function(result) {
                            videoKeyProvided = true;
                        }).catch(function(error) {
                            forceTestFailureFromPromise(test, error);
                        });
                    } else {
                        assert_equals(event.target, audioMediaKeySession);
                        assert_equals(String.fromCharCode.apply(null, keyId), audioKeyId);
                        var jwkSet = stringToUint8Array(createJWKSet(createJWK(keyId, audioKey)));
                        audioMediaKeySession.update(jwkSet).then(function(result) {
                            audioKeyProvided = true;
                        }).catch(function(error) {
                            forceTestFailureFromPromise(test, error);
                        });
                    }
                }

                function onPlaying(event)
                {
                    // Video should not start playing until both keys have been
                    // provided.
                    assert_true(videoKeyProvided);
                    assert_true(audioKeyProvided);

                    // Not using waitForEventAndRunStep() to avoid too many
                    // EVENT(onTimeUpdate) logs.
                    video.addEventListener('timeupdate', onTimeUpdate, true);
                }

                function onTimeUpdate(event)
                {
                    if (event.target.currentTime < 0.2)
                        return;

                    test.done();
                }

                MediaKeys.create('org.w3.clearkey').then(function(mediaKeys) {

                    waitForEventAndRunStep('needkey', video, onNeedKey, test);
                    waitForEventAndRunStep('playing', video, onPlaying, test);

                    video.setMediaKeys(mediaKeys);
                    video.src = '../content/test-encrypted-different-av-keys.webm';
                    video.play();
                }).catch(function(error) {
                    forceTestFailureFromPromise(test, error);
                });
            }, 'Playback using Clear Key key system with multiple sessions.');
        </script>
    </body>
</html>
